+++ litemall-admin-api/src/main/java/org/linlinjava/litemall/admin/web/AdminAddressController.java
@@ -17,8 +17,8 @@ import java.util.Map;
 
 @RestController
 @RequestMapping("/admin/address")
-public class AddressController {
-    private final Log logger = LogFactory.getLog(AddressController.class);
+public class AdminAddressController {
+    private final Log logger = LogFactory.getLog(AdminAddressController.class);
 
     @Autowired
     private LitemallAddressService addressService;
