+++ litemall-admin-api/src/main/java/org/linlinjava/litemall/admin/web/AdminGoodsAttributeController.java
@@ -3,8 +3,8 @@ package org.linlinjava.litemall.admin.web;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.linlinjava.litemall.admin.annotation.LoginAdmin;
-import org.linlinjava.litemall.db.domain.LitemallGoods;
-import org.linlinjava.litemall.db.service.LitemallGoodsService;
+import org.linlinjava.litemall.db.domain.LitemallGoodsAttribute;
+import org.linlinjava.litemall.db.service.LitemallGoodsAttributeService;
 import org.linlinjava.litemall.core.util.ResponseUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -14,16 +14,16 @@ import java.util.List;
 import java.util.Map;
 
 @RestController
-@RequestMapping("/admin/goods")
-public class GoodsController {
-    private final Log logger = LogFactory.getLog(GoodsController.class);
+@RequestMapping("/admin/goods-attribute")
+public class AdminGoodsAttributeController {
+    private final Log logger = LogFactory.getLog(AdminGoodsAttributeController.class);
 
     @Autowired
-    private LitemallGoodsService goodsService;
+    private LitemallGoodsAttributeService goodsAttributeService;
 
     @GetMapping("/list")
     public Object list(@LoginAdmin Integer adminId,
-                       String goodsSn, String name,
+                       Integer goodsId,
                        @RequestParam(value = "page", defaultValue = "1") Integer page,
                        @RequestParam(value = "limit", defaultValue = "10") Integer limit,
                        String sort, String order){
@@ -31,22 +31,22 @@ public class GoodsController {
             return ResponseUtil.unlogin();
         }
 
-        List<LitemallGoods> goodsList = goodsService.querySelective(goodsSn, name, page, limit, sort, order);
-        int total = goodsService.countSelective(goodsSn, name, page, limit, sort, order);
+        List<LitemallGoodsAttribute> goodsAttributeList = goodsAttributeService.querySelective(goodsId, page, limit, sort, order);
+        int total = goodsAttributeService.countSelective(goodsId, page, limit, sort, order);
         Map<String, Object> data = new HashMap<>();
         data.put("total", total);
-        data.put("items", goodsList);
+        data.put("items", goodsAttributeList);
 
         return ResponseUtil.ok(data);
     }
 
     @PostMapping("/create")
-    public Object create(@LoginAdmin Integer adminId, @RequestBody LitemallGoods goods){
+    public Object create(@LoginAdmin Integer adminId, @RequestBody LitemallGoodsAttribute goodsAttribute){
         if(adminId == null){
             return ResponseUtil.unlogin();
         }
-        goodsService.add(goods);
-        return ResponseUtil.ok(goods);
+        goodsAttributeService.add(goodsAttribute);
+        return ResponseUtil.ok(goodsAttribute);
     }
 
     @GetMapping("/read")
@@ -59,25 +59,25 @@ public class GoodsController {
             return ResponseUtil.badArgument();
         }
 
-        LitemallGoods goods = goodsService.findById(id);
-        return ResponseUtil.ok(goods);
+        LitemallGoodsAttribute goodsAttribute = goodsAttributeService.findById(id);
+        return ResponseUtil.ok(goodsAttribute);
     }
 
     @PostMapping("/update")
-    public Object update(@LoginAdmin Integer adminId, @RequestBody LitemallGoods goods){
+    public Object update(@LoginAdmin Integer adminId, @RequestBody LitemallGoodsAttribute goodsAttribute){
         if(adminId == null){
             return ResponseUtil.unlogin();
         }
-        goodsService.updateById(goods);
-        return ResponseUtil.ok(goods);
+        goodsAttributeService.updateById(goodsAttribute);
+        return ResponseUtil.ok(goodsAttribute);
     }
 
     @PostMapping("/delete")
-    public Object delete(@LoginAdmin Integer adminId, @RequestBody LitemallGoods goods){
+    public Object delete(@LoginAdmin Integer adminId, @RequestBody LitemallGoodsAttribute goodsAttribute){
         if(adminId == null){
             return ResponseUtil.unlogin();
         }
-        goodsService.deleteById(goods.getId());
+        goodsAttributeService.deleteById(goodsAttribute.getId());
         return ResponseUtil.ok();
     }
