+++ litemall-admin-api/src/main/java/org/linlinjava/litemall/admin/web/AdminGoodsController.java
@@ -3,8 +3,8 @@ package org.linlinjava.litemall.admin.web;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.linlinjava.litemall.admin.annotation.LoginAdmin;
-import org.linlinjava.litemall.db.domain.LitemallSearchHistory;
-import org.linlinjava.litemall.db.service.LitemallSearchHistoryService;
+import org.linlinjava.litemall.db.domain.LitemallGoods;
+import org.linlinjava.litemall.db.service.LitemallGoodsService;
 import org.linlinjava.litemall.core.util.ResponseUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -14,16 +14,16 @@ import java.util.List;
 import java.util.Map;
 
 @RestController
-@RequestMapping("/admin/history")
-public class HistoryController {
-    private final Log logger = LogFactory.getLog(HistoryController.class);
+@RequestMapping("/admin/goods")
+public class AdminGoodsController {
+    private final Log logger = LogFactory.getLog(AdminGoodsController.class);
 
     @Autowired
-    private LitemallSearchHistoryService searchHistoryService;
+    private LitemallGoodsService goodsService;
 
     @GetMapping("/list")
     public Object list(@LoginAdmin Integer adminId,
-                       String userId, String keyword,
+                       String goodsSn, String name,
                        @RequestParam(value = "page", defaultValue = "1") Integer page,
                        @RequestParam(value = "limit", defaultValue = "10") Integer limit,
                        String sort, String order){
@@ -31,21 +31,22 @@ public class HistoryController {
             return ResponseUtil.unlogin();
         }
 
-        List<LitemallSearchHistory> footprintList = searchHistoryService.querySelective(userId, keyword, page, limit, sort, order);
-        int total = searchHistoryService.countSelective(userId, keyword, page, limit, sort, order);
+        List<LitemallGoods> goodsList = goodsService.querySelective(goodsSn, name, page, limit, sort, order);
+        int total = goodsService.countSelective(goodsSn, name, page, limit, sort, order);
         Map<String, Object> data = new HashMap<>();
         data.put("total", total);
-        data.put("items", footprintList);
+        data.put("items", goodsList);
 
         return ResponseUtil.ok(data);
     }
 
     @PostMapping("/create")
-    public Object create(@LoginAdmin Integer adminId, @RequestBody LitemallSearchHistory history){
+    public Object create(@LoginAdmin Integer adminId, @RequestBody LitemallGoods goods){
         if(adminId == null){
-            return ResponseUtil.fail401();
+            return ResponseUtil.unlogin();
         }
-        return ResponseUtil.fail501();
+        goodsService.add(goods);
+        return ResponseUtil.ok(goods);
     }
 
     @GetMapping("/read")
@@ -58,25 +59,25 @@ public class HistoryController {
             return ResponseUtil.badArgument();
         }
 
-        LitemallSearchHistory history = searchHistoryService.findById(id);
-        return ResponseUtil.ok(history);
+        LitemallGoods goods = goodsService.findById(id);
+        return ResponseUtil.ok(goods);
     }
 
     @PostMapping("/update")
-    public Object update(@LoginAdmin Integer adminId, @RequestBody LitemallSearchHistory history){
+    public Object update(@LoginAdmin Integer adminId, @RequestBody LitemallGoods goods){
         if(adminId == null){
             return ResponseUtil.unlogin();
         }
-        searchHistoryService.updateById(history);
-        return ResponseUtil.ok();
+        goodsService.updateById(goods);
+        return ResponseUtil.ok(goods);
     }
 
     @PostMapping("/delete")
-    public Object delete(@LoginAdmin Integer adminId, @RequestBody LitemallSearchHistory history){
+    public Object delete(@LoginAdmin Integer adminId, @RequestBody LitemallGoods goods){
         if(adminId == null){
             return ResponseUtil.unlogin();
         }
-        searchHistoryService.deleteById(history.getId());
+        goodsService.deleteById(goods.getId());
         return ResponseUtil.ok();
     }
