+++ litemall-admin-api/src/main/java/org/linlinjava/litemall/admin/web/AdminKeywordController.java
@@ -3,8 +3,8 @@ package org.linlinjava.litemall.admin.web;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.linlinjava.litemall.admin.annotation.LoginAdmin;
-import org.linlinjava.litemall.db.domain.LitemallTopic;
-import org.linlinjava.litemall.db.service.LitemallTopicService;
+import org.linlinjava.litemall.db.domain.LitemallKeyword;
+import org.linlinjava.litemall.db.service.LitemallKeywordService;
 import org.linlinjava.litemall.core.util.ResponseUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -14,16 +14,16 @@ import java.util.List;
 import java.util.Map;
 
 @RestController
-@RequestMapping("/admin/topic")
-public class TopicController {
-    private final Log logger = LogFactory.getLog(TopicController.class);
+@RequestMapping("/admin/keyword")
+public class AdminKeywordController {
+    private final Log logger = LogFactory.getLog(AdminKeywordController.class);
 
     @Autowired
-    private LitemallTopicService topicService;
+    private LitemallKeywordService keywordService;
 
     @GetMapping("/list")
     public Object list(@LoginAdmin Integer adminId,
-                       String title, String subtitle,
+                       String keyword, String url,
                        @RequestParam(value = "page", defaultValue = "1") Integer page,
                        @RequestParam(value = "limit", defaultValue = "10") Integer limit,
                        String sort, String order){
@@ -31,22 +31,22 @@ public class TopicController {
             return ResponseUtil.unlogin();
         }
 
-        List<LitemallTopic> topicList = topicService.querySelective(title, subtitle, page, limit, sort, order);
-        int total = topicService.countSelective(title, subtitle, page, limit, sort, order);
+        List<LitemallKeyword> brandList = keywordService.querySelective(keyword, url, page, limit, sort, order);
+        int total = keywordService.countSelective(keyword, url, page, limit, sort, order);
         Map<String, Object> data = new HashMap<>();
         data.put("total", total);
-        data.put("items", topicList);
+        data.put("items", brandList);
 
         return ResponseUtil.ok(data);
     }
 
     @PostMapping("/create")
-    public Object create(@LoginAdmin Integer adminId, @RequestBody LitemallTopic topic){
+    public Object create(@LoginAdmin Integer adminId, @RequestBody LitemallKeyword keywords){
         if(adminId == null){
             return ResponseUtil.unlogin();
         }
-        topicService.add(topic);
-        return ResponseUtil.ok(topic);
+        keywordService.add(keywords);
+        return ResponseUtil.ok(keywords);
     }
 
     @GetMapping("/read")
@@ -59,25 +59,25 @@ public class TopicController {
             return ResponseUtil.badArgument();
         }
 
-        LitemallTopic brand = topicService.findById(id);
+        LitemallKeyword brand = keywordService.findById(id);
         return ResponseUtil.ok(brand);
     }
 
     @PostMapping("/update")
-    public Object update(@LoginAdmin Integer adminId, @RequestBody LitemallTopic topic){
+    public Object update(@LoginAdmin Integer adminId, @RequestBody LitemallKeyword keywords){
         if(adminId == null){
             return ResponseUtil.unlogin();
         }
-        topicService.updateById(topic);
-        return ResponseUtil.ok(topic);
+        keywordService.updateById(keywords);
+        return ResponseUtil.ok(keywords);
     }
 
     @PostMapping("/delete")
-    public Object delete(@LoginAdmin Integer adminId, @RequestBody LitemallTopic topic){
+    public Object delete(@LoginAdmin Integer adminId, @RequestBody LitemallKeyword brand){
         if(adminId == null){
             return ResponseUtil.unlogin();
         }
-        topicService.deleteById(topic.getId());
+        keywordService.deleteById(brand.getId());
         return ResponseUtil.ok();
     }
