+++ litemall-admin-api/src/main/java/org/linlinjava/litemall/admin/web/AdminTopicController.java
@@ -3,8 +3,8 @@ package org.linlinjava.litemall.admin.web;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.linlinjava.litemall.admin.annotation.LoginAdmin;
-import org.linlinjava.litemall.db.domain.LitemallFootprint;
-import org.linlinjava.litemall.db.service.LitemallFootprintService;
+import org.linlinjava.litemall.db.domain.LitemallTopic;
+import org.linlinjava.litemall.db.service.LitemallTopicService;
 import org.linlinjava.litemall.core.util.ResponseUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -14,16 +14,16 @@ import java.util.List;
 import java.util.Map;
 
 @RestController
-@RequestMapping("/admin/footprint")
-public class FootprintController {
-    private final Log logger = LogFactory.getLog(FootprintController.class);
+@RequestMapping("/admin/topic")
+public class AdminTopicController {
+    private final Log logger = LogFactory.getLog(AdminTopicController.class);
 
     @Autowired
-    private LitemallFootprintService footprintService;
+    private LitemallTopicService topicService;
 
     @GetMapping("/list")
     public Object list(@LoginAdmin Integer adminId,
-                       String userId, String goodsId,
+                       String title, String subtitle,
                        @RequestParam(value = "page", defaultValue = "1") Integer page,
                        @RequestParam(value = "limit", defaultValue = "10") Integer limit,
                        String sort, String order){
@@ -31,21 +31,22 @@ public class FootprintController {
             return ResponseUtil.unlogin();
         }
 
-        List<LitemallFootprint> footprintList = footprintService.querySelective(userId, goodsId, page, limit, sort, order);
-        int total = footprintService.countSelective(userId, goodsId, page, limit, sort, order);
+        List<LitemallTopic> topicList = topicService.querySelective(title, subtitle, page, limit, sort, order);
+        int total = topicService.countSelective(title, subtitle, page, limit, sort, order);
         Map<String, Object> data = new HashMap<>();
         data.put("total", total);
-        data.put("items", footprintList);
+        data.put("items", topicList);
 
         return ResponseUtil.ok(data);
     }
 
     @PostMapping("/create")
-    public Object create(@LoginAdmin Integer adminId, @RequestBody LitemallFootprint footprint){
+    public Object create(@LoginAdmin Integer adminId, @RequestBody LitemallTopic topic){
         if(adminId == null){
             return ResponseUtil.unlogin();
         }
-        return ResponseUtil.unsupport();
+        topicService.add(topic);
+        return ResponseUtil.ok(topic);
     }
 
     @GetMapping("/read")
@@ -58,25 +59,25 @@ public class FootprintController {
             return ResponseUtil.badArgument();
         }
 
-        LitemallFootprint footprint = footprintService.findById(id);
-        return ResponseUtil.ok(footprint);
+        LitemallTopic brand = topicService.findById(id);
+        return ResponseUtil.ok(brand);
     }
 
     @PostMapping("/update")
-    public Object update(@LoginAdmin Integer adminId, @RequestBody LitemallFootprint footprint){
+    public Object update(@LoginAdmin Integer adminId, @RequestBody LitemallTopic topic){
         if(adminId == null){
             return ResponseUtil.unlogin();
         }
-        footprintService.updateById(footprint);
-        return ResponseUtil.ok();
+        topicService.updateById(topic);
+        return ResponseUtil.ok(topic);
     }
 
     @PostMapping("/delete")
-    public Object delete(@LoginAdmin Integer adminId, @RequestBody LitemallFootprint footprint){
+    public Object delete(@LoginAdmin Integer adminId, @RequestBody LitemallTopic topic){
         if(adminId == null){
             return ResponseUtil.unlogin();
         }
-        footprintService.deleteById(footprint.getId());
+        topicService.deleteById(topic.getId());
         return ResponseUtil.ok();
     }
